auth_service/tests/functional/requirements.txt
aiohttp==3.8.6
backoff==2.2.1
fastapi==0.110.0
fastapi-pagination==0.12.24
pydantic==2.6.1
email-validator==2.1.1
pydantic-settings==2.0.2
gunicorn==21.2.0
uvicorn==0.27.0.post1
redis==5.0.1
orjson==3.9.13
sqlalchemy==2.0.28
alembic==1.11.1
psycopg2-binary==2.9.6
asyncpg==0.27.0
httpx==0.23.3
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-mock==3.11.1

# AUTH
async-fastapi-jwt-auth==0.6.3
passlib==1.7.4
werkzeug==3.0.1